task 1

+++ doubly_linked_lists/1-dlistint_len.c
@@ -6,11 +6,11 @@
 */
 size_t dlistint_len(const dlistint_t *h)
 {
-	unsigned int data = 0;
+	size_t data = 0;
 
 	while (h)
 		h = h->next;
-		data++;
+			data++;
 
 
 	return (data);
